Separated yaml and properties file source and added support for additional file configuration sources
